pyproject.toml
[project]
name = "tibame-tjr101-t1"
version = "0.1.0"
description = "The TJR101 Team 1 Project"
authors = [
    {name = "TJR101 Team1"}
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "selenium (>=4.31.0,<5.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "webdriver-manager (>=4.0.2,<5.0.0)",
    "geopy (>=2.4.1,<3.0.0)",
    "py-dotenv (>=0.1,<0.2)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "pymysql (>=1.1.1,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
